remove tests

+++ app/db/models/__init__.py
@@ -1,5 +1,5 @@
-####################
-# TODO to be deleted
+import os
+
 from app.db import get_db
 from app.db.models.dues_payments import DuesPayment
 from app.db.models.items import Item, Category
@@ -11,23 +11,23 @@ from app.db.models.seller_items import SellerItems
 from app.db.models.sellers import Seller, ExpenseAccount
 
 
-def clear_db():
-    _db = next(get_db())
-    try:
-        _db.query(MemberHistory).delete()
-        _db.query(MemberDonation).delete()
-        _db.query(MemberDuesPayment).delete()
+def clear_db() -> None:
+    if os.getenv("CECC_MODE") == "TEST":
+        _db = next(get_db())
+        try:
+            _db.query(MemberDonation).delete()
+            _db.query(MemberDuesPayment).delete()
+            _db.query(MemberHistory).delete()
 
-        _db.query(MemberItems).delete()
-        _db.query(SellerItems).delete()
+            _db.query(SellerItems).delete()
+            _db.query(MemberItems).delete()
+            _db.query(Item).delete()
 
-        _db.query(Item).delete()
-        _db.query(Category).delete()
-        _db.query(ExpenseAccount).delete()
-        _db.query(Seller).delete()
-        _db.query(DuesPayment).delete()
-        _db.query(Member).delete()
-        _db.commit()
-    except:
-        _db.rollback()
-####################
+            _db.query(DuesPayment).delete()
+            _db.query(Category).delete()
+            _db.query(ExpenseAccount).delete()
+            _db.query(Seller).delete()
+            _db.query(Member).delete()
+            _db.commit()
+        except:
+            _db.rollback()
